+++ integration/cmd/osrm-files-extractor/main.go
@@ -5,9 +5,9 @@ import (
 	"strings"
 
 	"github.com/Telenav/osrm-backend/integration/osrmfiles"
+	"github.com/Telenav/osrm-backend/integration/osrmfiles/dotnbgnodes"
 	"github.com/Telenav/osrm-backend/integration/osrmfiles/dotosrm"
-	"github.com/Telenav/osrm-backend/integration/osrmfiles/dotosrmdotnbgnodes"
-	"github.com/Telenav/osrm-backend/integration/osrmfiles/dotosrmdottimestamp"
+	"github.com/Telenav/osrm-backend/integration/osrmfiles/dottimestamp"
 
 	"github.com/golang/glog"
 )
@@ -27,8 +27,8 @@ func createEmptyOSRMFilesContents(osrmBasefilePath string) map[string]osrmfiles.
 
 	m := map[string]osrmfiles.ContentsOperator{}
 	m[dotOSRMSuffix] = dotosrm.New(osrmBasefilePath)
-	m[dotOSRMDotTimestampSuffix] = dotosrmdottimestamp.New(osrmBasefilePath + dotTimestampSuffix)
-	m[dotOSRMDotNBGNodesSuffix] = dotosrmdotnbgnodes.New(osrmBasefilePath + dotNBGNodesSuffix)
+	m[dotOSRMDotTimestampSuffix] = dottimestamp.New(osrmBasefilePath + dotTimestampSuffix)
+	m[dotOSRMDotNBGNodesSuffix] = dotnbgnodes.New(osrmBasefilePath + dotNBGNodesSuffix)
 
 	return m
 }
